pyproject.toml
[project]
name = "traintimes"
description = "A Python SDK for realtimetrains' API"
version = "0.1.0.dev0"
readme = "README.rst"
requires-python = ">=3.10"
authors = [
  { name = "Tom Viner", email = "traintimes@viner.tv" },
]
dependencies = [
  "purl>=1.1,<2",
  "requests>=2.31,<3",
  "requests-cache>=1.2,<2",
  "python-dotenv>=1,<2",
]

[project.urls]
Homepage = "https://github.com/tomviner/traintimes"

[dependency-groups]
test = [
  "freezegun==0.3.5",
  "pytest",
  "pytest-cov",
  "requests-mock",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.ruff]
target-version = "py310"
force-exclude = true
exclude = [".venv"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle Error
  "W",  # pycodestyle Warning
  "F",  # Pyflakes
  "I",  # isort
  "C90",  # mccabe (code complexity)
  "T10",  # flake8-debugger
]

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.coverage.run]
branch = true
source = ["traintimes"]

[tool.coverage.report]
show_missing = true
skip_covered = true
precision = 1
fail_under = 100

[tool.pytest.ini_options]
addopts = "--cov"
